Rename files and references
