+++ packages/core-data/src/fetch/fetch-link-suggestions.js
