+++ packages/core-data/src/fetch/index.js
@@ -1,2 +1,2 @@
-export { default as fetchLinkSuggestions } from './__experimental-fetch-link-suggestions';
+export { default as fetchLinkSuggestions } from './fetch-link-suggestions';
 export { default as __experimentalFetchUrlData } from './__experimental-fetch-url-data';
